+++ projects/start-os/web/ui/src/app/services/gateway.service.ts
@@ -2,6 +2,7 @@ import { inject, Injectable } from '@angular/core'
 import { PatchDB } from 'patch-db-client'
 import { T, utils } from '@start9labs/start-core'
 import { map } from 'rxjs'
+import { getWanIp } from '../utils/gua'
 import { DataModel } from './patch-db/data-model'
 import { toSignal } from '@angular/core/rxjs-interop'
 
@@ -11,7 +12,8 @@ export type GatewayPlus = T.NetworkInterfaceInfo & {
   ipInfo: T.IpInfo
   subnets: utils.IpNet[]
   lanIpv4: string[]
-  wanIp?: utils.IpAddress
+  /** The pinned override when set, else what StartOS detected. */
+  effectiveWanIp: string | null
 }
 
 @Injectable()
@@ -45,8 +47,7 @@ export class GatewayService {
               name,
               subnets,
               lanIpv4: subnets.filter(s => s.isIpv4()).map(s => s.address),
-              wanIp:
-                val.ipInfo?.wanIp && utils.IpAddress.parse(val.ipInfo?.wanIp),
+              effectiveWanIp: getWanIp(val),
             } as GatewayPlus
           })
       }),